Renames convertFromHTMLToContentBlocks2 to convertFromHTMLToContentBlocks
