+++ src/Draft.js
@@ -36,7 +36,7 @@ const getDefaultKeyBinding = require('getDefaultKeyBinding');
 const getVisibleSelectionRect = require('getVisibleSelectionRect');
 const gkx = require('gkx');
 
-const convertFromHTML = require('convertFromHTMLToContentBlocks2');
+const convertFromHTML = require('convertFromHTMLToContentBlocks');
 
 const DraftPublic = {
   Editor: DraftEditor,
